backend/requirements.txt.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-filter>=23.0
psycopg2-binary>=2.9
gunicorn>=21.2
python-dotenv>=1.0
django-cors-headers>=4.0
drf-spectacular>=0.26  # Para documentación Swagger
pytest>=7.4
pytest-django>=4.5
ruff>=0.1
black>=23.0